app/requirements.txt
python-dotenv==1.1.0
Flask==3.0.3
Flask-Cors==6.0.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
gunicorn==22.0.0
requests==2.32.3
urllib3==1.26.19
marshmallow==4.0.1